requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
distro==1.9.0
fastapi==0.115.12
gitdb==4.0.12
GitPython==3.1.44
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
narwhals==1.41.0
numpy==2.2.5
openai==1.77.0
packaging==24.2
pandas==2.2.3
pillow==11.2.1
praw==7.8.1
prawcore==2.4.0
producthunt==0.1.4
protobuf==6.31.0
pyarrow==20.0.0
pydantic==2.11.4
pydantic_core==2.33.2
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.25.1
setuptools==78.1.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.46.2
streamlit==1.45.1
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
update-checker==0.18.0
urllib3==2.4.0
uvicorn==0.34.2
watchdog==6.0.0
websocket-client==1.8.0
wheel==0.45.1
